moves utils into distinct packages
