+++ utils/system/system.go
@@ -1,4 +1,4 @@
-package utils
+package system
 
 import (
 	"errors"
